Did more rearranging of the component structure for the navbar and added in some new button components.
