+++ client/src/index.css
@@ -52,11 +52,6 @@ html, body, #root, .App {
   margin: auto;
 }
 
-.logo {
-  color: brown;
-  font-size: 20px;
-}
-
 form {
   display: flex;
   flex-direction: column;
